refactor(gov): move governance query methods to query_group
Moved all Conway governance query-related methods from the dedicated
`gov_query_group.py` module into the existing `query_group.py` as
methods with a `get_` prefix. Removed the now-redundant
`gov_query_group.py` and its references from `gov_group.py`. This
consolidates governance query functionality under a single group,
simplifying the codebase and reducing duplication.
